DDP-5045 - fix PR comments

+++ pepper-apis/src/main/java/org/broadinstitute/ddp/transformers/UnixTimestampToInstantAdapter.java
@@ -29,6 +29,6 @@ public class UnixTimestampToInstantAdapter extends TypeAdapter<Instant> {
             in.nextNull();
             return null;
         }
-        return Instant.ofEpochSecond(in.nextInt());
+        return Instant.ofEpochSecond(in.nextLong());
     }
 }
